- Refactored the codebase and move code to their respective new libraries
