+++ IQKeyboardManagerSwift/UIKitExtensions/UIView+Parent.swift
@@ -87,6 +87,4 @@ public extension IQKeyboardManagerWrapper where Base: UIView {
             return parentContainerViewController
         }
     }
-
-
 }
